+++ pipda/function.py
@@ -65,15 +65,27 @@ class FunctionCall(Expression):
                 },
             )
 
-        bound = func.bind_arguments(*self._pipda_args, **self._pipda_kwargs)
         context = func.contexts["_"] or context
-        extra_contexts = func.extra_contexts
-        for key, val in bound.arguments.items():
-            ctx = extra_contexts["_"].get(key, context)
-            val = evaluate_expr(val, data, ctx)
-            bound.arguments[key] = val
+        extra_contexts = func.extra_contexts["_"]
 
-        return func.func(*bound.args, **bound.kwargs)
+        if extra_contexts:
+            bound = func.bind_arguments(*self._pipda_args, **self._pipda_kwargs)
+
+            for key, val in bound.arguments.items():
+                ctx = extra_contexts.get(key, context)
+                val = evaluate_expr(val, data, ctx)
+                bound.arguments[key] = val
+
+            return func.func(*bound.args, **bound.kwargs)
+
+        # we don't need signature if there is no extra context
+        return func.func(
+            *(evaluate_expr(arg, data, context) for arg in self._pipda_args),
+            **{
+                key: evaluate_expr(val, data, context)
+                for key, val in self._pipda_kwargs.items()
+            },
+        )
 
 
 class Registered(ABC):